Respond to write cancellation by closing output
Close an output stream registered to a WriteResponseObserver when a
request is cancelled. This should prevent hanging outputs with no hope
of closure, which should consequently prevent observed deadlocks with
write caching in CASFileCache capable of stalling unrelated
service queryWriteStatus/writes, as well as critical worker
functionality like input fetch and report result.